api/requirements.txt
psycopg[c]>=3.2.6,<3.3
Django>=5.2,<5.3
djangorestframework>=3.15.2,<3.16
djangorestframework-gis>=1.1,<1.2
django-vectortiles>=1.0.1,<1.1
django-silk>=5.3.2,<5.4
drf-spectacular>=0.28.0,<0.29
pytest>=8.3.5,<8.4
pytest-django>=4.11.1,<4.12

# Others
geopandas>=1.0.1,<1.1